+++ domain/mocks/orderbook_repository_mock.go
@@ -12,6 +12,8 @@ type OrderbookRepositoryMock struct {
 	GetAllTicksFunc func(poolID uint64) (map[int64]orderbookdomain.OrderbookTick, bool)
 	GetTicksFunc    func(poolID uint64, tickIDs []int64) (map[int64]orderbookdomain.OrderbookTick, error)
 	GetTickByIDFunc func(poolID uint64, tickID int64) (orderbookdomain.OrderbookTick, bool)
+	StoreOrdersFunc func(poolID uint64, orders []orderbookdomain.Order) error
+	GetOrdersFunc   func(poolID uint64, ownerAddress string) ([]orderbookdomain.Order, bool)
 }
 
 // StoreTicks implements OrderBookRepository.
@@ -51,4 +53,32 @@ func (m *OrderbookRepositoryMock) GetTickByID(poolID uint64, tickID int64) (orde
 		return m.GetTickByIDFunc(poolID, tickID)
 	}
 	panic("GetTickByID not implemented")
+}
+
+func (m *OrderbookRepositoryMock) WithStoreOrders(err error) {
+	m.StoreOrdersFunc = func(poolID uint64, orders []orderbookdomain.Order) error {
+		return err
+	}
+}
+
+// StoreOrders implements OrderBookRepository.
+func (m *OrderbookRepositoryMock) StoreOrders(poolID uint64, orders []orderbookdomain.Order) error {
+	if m.StoreOrdersFunc != nil {
+		return m.StoreOrdersFunc(poolID, orders)
+	}
+	panic("StoreOrders not implemented")
+}
+
+func (m *OrderbookRepositoryMock) WithGetOrdersFunc(orders []orderbookdomain.Order, ok bool) {
+	m.GetOrdersFunc = func(poolID uint64, ownerAddress string) ([]orderbookdomain.Order, bool) {
+		return orders, ok
+	}
+}
+
+// GetOrders implements OrderBookRepository.
+func (m *OrderbookRepositoryMock) GetOrders(poolID uint64, ownerAddress string) ([]orderbookdomain.Order, bool) {
+	if m.GetOrdersFunc != nil {
+		return m.GetOrdersFunc(poolID, ownerAddress)
+	}
+	panic("GetOrders not implemented")
 }