+++ domain/orderbook/plugin/config.go
@@ -2,7 +2,8 @@ package orderbookplugindomain
 
 // Orderbook plugin names
 const (
-	OrderbookFillbotPlugin  = "orderbook-fillbot-plugin"
-	OrderbookClaimbotPlugin = "orderbook-claimbot-plugin"
-	CustomSubmodulePlugin   = "custom-submodule-plugin"
+	OrderbookFillbotPlugin     = "orderbook-fillbot-plugin"
+	OrderbookClaimbotPlugin    = "orderbook-claimbot-plugin"
+	OrderbookOrdersCachePlugin = "orderbook-orders-cache-plugin"
+	CustomSubmodulePlugin      = "custom-submodule-plugin"
 )