+++ ingest/usecase/plugins/orderbook/orderscache/export_test.go
@@ -1,0 +1,10 @@
+package orderbookorderscache
+
+import (
+	"github.com/osmosis-labs/sqs/domain"
+	"github.com/osmosis-labs/sqs/domain/mvc"
+)
+
+func GetOrderbooks(poolsUsecase mvc.PoolsUsecase, metadata domain.BlockPoolMetadata) ([]domain.CanonicalOrderBooksResult, error) {
+	return getOrderbooks(poolsUsecase, metadata)
+}