+++ CHANGELOG.md
@@ -35,6 +35,10 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 # Changelog
 
+## Unreleased
+
+- #622 - feat: orderbook orders cache plugin
+
 ## v28.3.1
 
 - #637 - BE-714 | Optimize route finding algorithm
@@ -42,7 +46,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ## v28.3.0
 
 - Remove dependency on keyring, wire in CosmosSigner and grab private key from SQS_PRIVATE_KEY environment variable.
-- #632 - E-697 | Increase max routes for pricing worker
+- #632 - BE-697 | Increase max routes for pricing worker
 - #633 - BE-707 | Skip processing quote requests when system is not healthy 
 
 ## v28.2.0